pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "cafa6"
version = "0.1.0"
description = "CAFA 6 Protein Function Prediction - Kaggle Competition"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.24",
    "pandas>=2.0",
    "scipy>=1.11",
    "scikit-learn>=1.3",
    "biopython>=1.81",
    "obonet>=1.0",
    "networkx>=3.1",
    "pyyaml>=6.0",
    "h5py>=3.9",
    "tqdm>=4.65",
    "torch>=2.1",
    "fair-esm>=2.0",
    "lightgbm>=4.0",
    "joblib>=1.3",
]

[project.optional-dependencies]
notebooks = [
    "jupyterlab>=4.0",
    "ipywidgets>=8.0",
    "matplotlib>=3.7",
    "seaborn>=0.12",
    "plotly>=5.15",
]
dev = [
    "ruff>=0.1",
    "pytest>=7.4",
]

[project.scripts]
cafa6-embed = "cafa6.embeddings:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
